# Conflicts:
#	src/main/java/ee/coffee/coffeeshop/controllers/OrderController.java
#	src/main/java/ee/coffee/coffeeshop/models/enums/OrderStatus.java
#	src/main/resources/templates/products.ftlh

+++ src/main/java/ee/coffee/coffeeshop/controllers/OrderController.java
@@ -33,12 +33,12 @@ public class OrderController {
     }
 
     @GetMapping(value = "/admin/order/{id}")
-    public Order getOrderById(@PathVariable(name = "id") Integer id) {
+    public Order getOrderById(@PathVariable(name = "id") Long id) {
         return orderService.getOrderById(id);
     }
 
     @PutMapping(value = "/admin/order/{id}/set_status/{status}")
-    public void setOrderStatus(@PathVariable(name = "status") OrderStatus status, @PathVariable(name = "id") Integer id) {
+    public void setOrderStatus(@PathVariable(name = "status") OrderStatus status, @PathVariable(name = "id") Long id) {
         orderService.setOrderStatus(status, id);
     }
 }
