+++ src/main/java/ee/coffee/coffeeshop/entity/Delivery.java
@@ -14,7 +14,7 @@ public class Delivery {
         @Id
         @GeneratedValue(strategy = GenerationType.IDENTITY)
         @Column(name = "delivery_id")
-        private Integer delivery_id;
+        private Long delivery_id;
 
         @Column(name = "delivery_status")
         private String delivery_status;
